Moved hook to hooks file
