+++ src/client/pages/HLI/HLI.hooks.tsx
@@ -1,4 +1,5 @@
 import { atom, useRecoilState } from 'recoil';
+import useSWRMutation from 'swr/dist/mutation';
 
 import { StadspasFrontend } from '../../../server/services/hli/stadspas-types';
 import { useAppStateGetter } from '../../hooks/useAppState';
@@ -28,4 +29,29 @@ export function useStadspassen() {
   );
 
   return [stadspassen, setStadspassenActiefStatus] as const;
+}
+
+export function useBlockStadspas(url: string | null, stadspasId: string) {
+  const setStadspassenActiefStatus = useStadspassen()[1];
+
+  return useSWRMutation(
+    url,
+    async (url) => {
+      const response = await fetch(url, {
+        method: 'POST',
+        credentials: 'include',
+      });
+
+      if (!response.ok) {
+        throw new Error('Request returned with an error');
+      }
+
+      setStadspassenActiefStatus((stadspasActiefState) => {
+        return { ...stadspasActiefState, [stadspasId]: false };
+      });
+
+      return response;
+    },
+    { revalidate: false, populateCache: false }
+  );
 }
